Fix unaligned access on filesystem object
The memory allocated to store the
objects is a char array which is
not garanteed to be 4-byte aligned.

Which means when you use this memory
to create new objects, the objects can
fall on unaligned addresses. Then
when you try to access the members
in this object, you end up with an
unaligned memory access fault.

This patch uses MBED_ALIGN directive
to force alignment